+++ pom.xml
@@ -35,7 +35,7 @@
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <maven.version>3.2.3</maven.version>
+    <maven.version>3.3.9</maven.version>
   </properties>
 
   <dependencies>
