+++ samples/data-jolokia-demo/pom.xml
@@ -22,12 +22,12 @@
   <parent>
     <groupId>io.fabric8</groupId>
     <artifactId>dmp-sample-parent</artifactId>
-    <version>0.19.0</version>
+    <version>0.19-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
 
   <artifactId>docker-jolokia-demo</artifactId>
-  <version>0.19.0</version>
+  <version>0.19-SNAPSHOT</version>
   <!-- add custom lifecycle -->
   <packaging>docker</packaging>
