+++ samples/docker-compose-demo/pom.xml
@@ -22,12 +22,12 @@
   <parent>
     <groupId>io.fabric8</groupId>
     <artifactId>dmp-sample-parent</artifactId>
-    <version>0.19.0</version>
+    <version>0.19-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
 
   <artifactId>docker-compose-demo</artifactId>
-  <version>0.19.0</version>
+  <version>0.19-SNAPSHOT</version>
 
   <url>http://www.jolokia.org</url>
