+++ samples/dockerignore/pom.xml
@@ -10,12 +10,12 @@
   <parent>
     <groupId>io.fabric8</groupId>
     <artifactId>dmp-sample-parent</artifactId>
-    <version>0.19.0</version>
+    <version>0.19-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
 
   <artifactId>dmp-sample-dockerignore</artifactId>
-  <version>0.19.0</version>
+  <version>0.19-SNAPSHOT</version>
   <packaging>docker-build</packaging>
 
   <build>
