+++ samples/healthcheck/pom.xml
@@ -12,12 +12,12 @@
   <parent>
     <groupId>io.fabric8</groupId>
     <artifactId>dmp-sample-parent</artifactId>
-    <version>0.19.0</version>
+    <version>0.19-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
 
   <artifactId>dmp-sample-healthcheck</artifactId>
-  <version>0.19.0</version>
+  <version>0.19-SNAPSHOT</version>
 
 
   <build>
