+++ samples/helloworld/src/main/java/io/fabric8/dmp/sample/helloworld/App.java
@@ -1,0 +1,10 @@
+package io.fabric8.dmp.sample.helloworld;
+
+/**
+ * Hello world!
+ */
+public class App {
+    public static void main(String[] args) {
+        System.out.println("Hello World!");
+    }
+}
