+++ samples/volume/pom.xml
@@ -15,12 +15,12 @@
   <parent>
     <groupId>io.fabric8</groupId>
     <artifactId>dmp-sample-parent</artifactId>
-    <version>0.19.0</version>
+    <version>0.19-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
 
   <artifactId>volume</artifactId>
-  <version>0.19.0</version>
+  <version>0.19-SNAPSHOT</version>
 
   <build>
     <plugins>
